Bugfix inicializacion de lista CPU EN ESPERA DE TCB

+++ kernel/src/conexiones.c
@@ -217,7 +217,7 @@ int32_t _procesar_conexion_nuevo_programa(char* codigo_beso, uint32_t longitud, 
 	// Agregamos la conexion a la lista de procesos
 	_agregar_conexion_a_procesos(conexion, pid);
 
-	conexion_consola(pid);
+	conexion_consola((uint32_t)pid);
 
 	return 0;
 }
